[REF] change name to more intuitive
